Add hardcoded gas prices as a fee component parameter
