+++ crates/core/component/fee/src/genesis.rs
@@ -1,20 +1,20 @@
 use anyhow::Context;
-use penumbra_proto::{penumbra::core::component::dao::v1alpha1 as pb, DomainType, TypeUrl};
+use penumbra_proto::{penumbra::core::component::fee::v1alpha1 as pb, DomainType, TypeUrl};
 use serde::{Deserialize, Serialize};
 
-use crate::params::DaoParameters;
+use crate::params::FeeParameters;
 
 #[derive(Deserialize, Serialize, Debug, Clone, Default)]
 #[serde(try_from = "pb::GenesisContent", into = "pb::GenesisContent")]
 pub struct Content {
-    /// The initial configuration parameters for the DAO component.
-    pub dao_params: DaoParameters,
+    /// The initial configuration parameters for the fee component.
+    pub fee_params: FeeParameters,
 }
 
 impl From<Content> for pb::GenesisContent {
     fn from(value: Content) -> Self {
         pb::GenesisContent {
-            dao_params: Some(value.dao_params.into()),
+            fee_params: Some(value.fee_params.into()),
         }
     }
 }
@@ -24,16 +24,16 @@ impl TryFrom<pb::GenesisContent> for Content {
 
     fn try_from(msg: pb::GenesisContent) -> Result<Self, Self::Error> {
         Ok(Content {
-            dao_params: msg
-                .dao_params
-                .context("dao params not present in protobuf message")?
+            fee_params: msg
+                .fee_params
+                .context("fee params not present in protobuf message")?
                 .try_into()?,
         })
     }
 }
 
 impl TypeUrl for Content {
-    const TYPE_URL: &'static str = "/penumbra.core.component.dao.v1alpha1.GenesisContent";
+    const TYPE_URL: &'static str = "/penumbra.core.component.fee.v1alpha1.GenesisContent";
 }
 
 impl DomainType for Content {
