+++ crates/core/transaction/src/gas.rs
@@ -1,5 +1,3 @@
-use std::ops::Add;
-
 use penumbra_chain::params::ChainParameters;
 use penumbra_compact_block::StatePayload;
 use penumbra_dao::{DaoDeposit, DaoOutput, DaoSpend};
@@ -7,9 +5,8 @@ use penumbra_dex::{
     BatchSwapOutputData, PositionClose, PositionOpen, PositionRewardClaim, PositionWithdraw, Swap,
     SwapClaim,
 };
+use penumbra_fee::Gas;
 use penumbra_ibc::{IbcAction, Ics20Withdrawal};
-use penumbra_num::Amount;
-use penumbra_proto::{core::transaction::v1alpha1 as pb, DomainType, TypeUrl};
 use penumbra_sct::Nullifier;
 use penumbra_shielded_pool::{Output, Spend};
 use penumbra_stake::{
@@ -22,29 +19,6 @@ use penumbra_governance::{
 };
 
 use crate::Action;
-
-/// Represents the different resources that a transaction can consume,
-/// for purposes of calculating multidimensional fees based on real
-/// transaction resource consumption.
-pub struct Gas {
-    pub block_space: u64,
-    pub compact_block_space: u64,
-    pub verification: u64,
-    pub execution: u64,
-}
-
-impl Add for Gas {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        Self {
-            block_space: self.block_space + rhs.block_space,
-            compact_block_space: self.compact_block_space + rhs.compact_block_space,
-            verification: self.verification + rhs.verification,
-            execution: self.execution + rhs.execution,
-        }
-    }
-}
 
 /// Allows [`Action`]s and [`Transaction`]s to statically indicate their relative resource consumption.
 /// Since the gas cost needs to be multiplied by a price, the values returned
@@ -495,57 +469,5 @@ impl GasCost for ValidatorDefinition {
             // Execution cost is currently hardcoded at 10 for all Action variants.
             execution: 10,
         }
-    }
-}
-
-/// Expresses the price of each unit of gas in terms of the staking token.
-#[derive(Clone, Debug)]
-pub struct GasPrices {
-    pub block_space_price: u64,
-    pub compact_block_space_price: u64,
-    pub verification_price: u64,
-    pub execution_price: u64,
-}
-
-impl GasPrices {
-    pub fn price(&self, gas: &Gas) -> Amount {
-        Amount::from(
-            self.block_space_price * gas.block_space
-                + self.compact_block_space_price * gas.compact_block_space
-                + self.verification_price * gas.verification
-                + self.execution_price * gas.execution,
-        )
-    }
-}
-
-impl TypeUrl for GasPrices {
-    const TYPE_URL: &'static str = "/penumbra.core.transaction.v1alpha1.GasPrices";
-}
-
-impl DomainType for GasPrices {
-    type Proto = pb::GasPrices;
-}
-
-impl From<GasPrices> for pb::GasPrices {
-    fn from(prices: GasPrices) -> Self {
-        pb::GasPrices {
-            block_space_price: prices.block_space_price,
-            compact_block_space_price: prices.compact_block_space_price,
-            verification_price: prices.verification_price,
-            execution_price: prices.execution_price,
-        }
-    }
-}
-
-impl TryFrom<pb::GasPrices> for GasPrices {
-    type Error = anyhow::Error;
-
-    fn try_from(proto: pb::GasPrices) -> Result<Self, Self::Error> {
-        Ok(GasPrices {
-            block_space_price: proto.block_space_price,
-            compact_block_space_price: proto.compact_block_space_price,
-            verification_price: proto.verification_price,
-            execution_price: proto.execution_price,
-        })
     }
 }
