+++ crates/proto/src/gen/penumbra.core.transaction.v1alpha1.rs
@@ -553,19 +553,3 @@ pub mod memo_view {
         Opaque(Opaque),
     }
 }
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct GasPrices {
-    /// The price per unit block space in terms of the staking token.
-    #[prost(uint64, tag = "1")]
-    pub block_space_price: u64,
-    /// The price per unit compact block space in terms of the staking token.
-    #[prost(uint64, tag = "2")]
-    pub compact_block_space_price: u64,
-    /// The price per unit verification cost in terms of the staking token.
-    #[prost(uint64, tag = "3")]
-    pub verification_price: u64,
-    /// The price per unit execution cost in terms of the staking token.
-    #[prost(uint64, tag = "4")]
-    pub execution_price: u64,
-}
