+++ proto/penumbra/penumbra/core/component/fee/v1alpha1/fee.proto
@@ -12,3 +12,26 @@ message Fee {
     // If absent, specifies the staking token implicitly.
     asset.v1alpha1.AssetId asset_id = 2;
 }
+
+message GasPrices {
+    // The price per unit block space in terms of the staking token.
+    uint64 block_space_price = 1;
+    // The price per unit compact block space in terms of the staking token.
+    uint64 compact_block_space_price = 2;
+    // The price per unit verification cost in terms of the staking token.
+    uint64 verification_price = 3;
+    // The price per unit execution cost in terms of the staking token.
+    uint64 execution_price = 4;
+}
+
+// Fee component configuration data.
+message FeeParameters {
+  // The current gas prices per resource.
+  GasPrices gas_prices = 1;
+}
+
+// Fee-specific genesis content.
+message GenesisContent {
+  // The FeeParameters present at genesis.
+  FeeParameters fee_params = 1;
+}
